Updated Movement and Jump with Velocity
Work-in-progress effort for satisfying movement:
Updated the movement to be velocity-based (to avoid clipping issues).

Tweaked a ground material for correcting speed.

Currently the rover drives with Tank controls on ground and mid-air.

Experimented with multiple movement calculations and rotation types but commented most of them out.

Experimented with how rotation/movement mid-air could be based on angle of camera (commented out).